[dom] Extract text nodes to their own module.
