+++ dom/src/lib.rs
@@ -12,6 +12,7 @@ pub mod elements;
 pub mod embed;
 pub mod interfaces;
 pub(crate) mod memo_node;
+pub mod text;
 
 /// A module for glob-importing the most commonly used moxie-dom items.
 pub mod prelude {
@@ -34,7 +35,7 @@ pub mod prelude {
             html_element::HtmlElement,
             node::{Node, Parent},
         },
-        memo_node::text,
+        text::text,
     };
 }
